chore: Move api utils around
