+++ mobile/src/api/album.ts
@@ -7,11 +7,11 @@ import { albums } from "~/db/schema";
 import i18next from "~/modules/i18n";
 
 import { iAsc } from "~/lib/drizzle";
-import type { DrizzleFilter } from "./types";
 import type {
+  DrizzleFilter,
   QueryManyWithTracksResult,
   QueryOneWithTracksResult,
-} from "./utils";
+} from "./types";
 import { getColumns } from "./utils";
 
 //#region GET Methods
